Module tardigrade::channel

Message channels for workflows.

Channels have a similar interface to async channels from the futures crate; their ends implement [Stream] / [Sink] traits. A workflow owns a single end of a channel (either a Receiver or Senders), while the other end is owned by the host environment or another workflow.

A Sender or Receiver can be obtained from the environment using WithHandle trait. This process is usually automated via workflow types. When executed in WASM, channels are backed by imported functions from the Tardigrade runtime. This is emulated for the test environment.

Structs

BroadcastError

Errors that can occur when sending a message over BroadcastPublisher.

BroadcastPublisher

[Sink] that publishes its values to zero or more BroadcastSubscribers.

BroadcastSubscriber

Subscriber [Stream] corresponding to a BroadcastPublisher.

Receiver

Receiver for a channel provided to the workflow.

RequestHandles

A pair consisting of a sender and a receiver that can be used to generate Requests.

Requests

Request sender based on a pair of channels: an outbound channel to send requests, and the corresponding inbound channel to listen to responses. Can be used to call to external task executors.

RequestsBuilder

Builder for Requests.

Response

Container for the response to a Request.

Sender

Sender for a workflow channel provided to the workflow.

Enums

Request

Container for a request that can also be used to cancel the pending request from the client side.

SendError

Errors that can occur when sending a message over a channel.

Functions

channel

Creates a new channel and returns its sender and receiver halves.

Type Definitions

RawReceiver

Receiver of raw byte messages.

RawSender

Sender of raw byte messages.