Enum tardigrade::channel::Request

pub enum Request<T> {
    New {
        id: u64,
        data: T,
        response_channel_id: ChannelId,
    },
    Cancel {
        id: u64,
    },
}

Container for a request that can also be used to cancel the pending request from the client side.

Variants

New

Fields

id: u64

Identifier of the request unique within the Requests instance generating requests.

data: T

Payload of the request.

response_channel_id: ChannelId

ID of the response channel.

New request.

Cancel

Fields

id: u64

Identifier of a previously created request.

Request cancellation.