Struct tardigrade::channel::BroadcastPublisher

pub struct BroadcastPublisher<T: Clone> { /* private fields */ }

[Sink] that publishes its values to zero or more BroadcastSubscribers.

The subscriber set is dynamic; subscribers can be created using Self::subscribe(), Cloned, and dropped. Each subscriber receives all values from the sink in the order they are sent to the sink.

To deal with lagging subscribers, publisher has an internal buffer of items with the capacity specified during creation. Once a subscriber is sufficiently behind, the publisher will not accept new items awaiting until the offending subscriber catches up or is dropped.

Implementations

impl<T: Clone> BroadcastPublisher<T>

pub fn new(capacity: usize) -> Self

Creates a new broadcast with the specified capacity.

Panics

Panics if capacity is zero.

pub fn subscribe(&self) -> BroadcastSubscriber<T>

Creates a new subscriber for this broadcast.