Struct tardigrade::channel::RequestHandles

pub struct RequestHandles<Req, Resp, C, Fmt: HandleFormat = Wasm>where
    C: Codec<Request<Req>> + Codec<Response<Resp>>,{
    pub requests: InEnv<Sender<Request<Req>, C>, Fmt>,
    pub responses: InEnv<Receiver<Response<Resp>, C>, Fmt>,
}

A pair consisting of a sender and a receiver that can be used to generate Requests.

Examples

RequestHandles can be used as a higher-level alternative to manually creating Requests.

#[derive(Serialize, Deserialize)]
pub struct MyRequest {
    // request fields...
}

#[derive(WithHandle)]
#[tardigrade(derive(Debug))]
pub struct MyWorkflow<Fmt: HandleFormat = Wasm> {
    pub task: RequestHandles<MyRequest, (), Json, Fmt>,
    // other handles...
}

// Usage in workflow code:
let handle: MyWorkflow = // ...
let (requests, _) =
    handle.task.process_requests().with_capacity(4).build();
requests.request(MyRequest { /* .. */ }).await?;

Fields
requests: InEnv<Sender<Request<Req>, C>, Fmt>

Requests sender. Will have path requests relative to the RequestHandles path.

responses: InEnv<Receiver<Response<Resp>, C>, Fmt>

Responses receiver. Will have path responses relative to the RequestHandles path.

Implementations

impl<Req: 'static, Resp: 'static, C> RequestHandles<Req, Resp, C>where
C: Codec<Request<Req>> + Codec<Response<Resp>>,

pub fn process_requests(self) -> RequestsBuilder<'static, Req, Resp, C>

Starts building a Requests processor based on the contained sender-receiver pair.