Struct tardigrade::channel::Requests

pub struct Requests<Req, Resp> { /* private fields */ }

Request sender based on a pair of channels: an outbound channel to send requests, and the corresponding inbound channel to listen to responses. Can be used to call a worker; for this, the outbound channel must specify the worker name.

Examples

A Requests instance can be built from a pair of sender / receiver channel halves:

#[derive(Debug, Serialize, Deserialize)]
pub struct MyRequest {
    // request fields...
}
#[derive(Debug, Serialize, Deserialize)]
pub struct MyResponse {
    // response fields...
}

#[derive(WithHandle, GetInterface)]
#[tardigrade(derive(Debug), auto_interface)]
pub struct MyWorkflow<Fmt: HandleFormat = Wasm> {
    pub requests: InEnv<Sender<Request<MyRequest>, Json>, Fmt>,
    pub responses: InEnv<Receiver<Response<MyResponse>, Json>, Fmt>,
}

#[async_trait(?Send)]
impl SpawnWorkflow for MyWorkflow {
    async fn spawn(_args: (), handle: Self) -> TaskResult {
        let (requests, _) = Requests::builder(handle.requests, handle.responses)
            .with_capacity(4)
            .with_task_name("handling_requests")
            .build();
        let response = requests
            .request(MyRequest { /* ... */ })
            .await
            .context("request cancelled")?;
        // Do something with the response...
    }
}

See RequestHandles docs for a higher-level alternative way to build Requests.

Implementations

impl<Req: 'static, Resp: 'static> Requests<Req, Resp>

pub fn builder<C>( requests_sx: Sender<Request<Req>, C>, responses_rx: Receiver<Response<Resp>, C> ) -> RequestsBuilder<'static, Req, Resp, C>where C: Codec<Request<Req>> + Codec<Response<Resp>>,

Creates a new sender based on the specified channels.

pub fn request( &self, request: Req ) -> impl Future<Output = Result<Resp, Canceled>>

Performs a request and returns a future that resolves to the response.

If the future is dropped before the response arrives, a cancellation signal will be sent to the external executor.

Errors

The returned future resolves to an error if the responses channel is closed before a response arrives.