Struct tardigrade::channel::Sender

pub struct Sender<T, C> { /* private fields */ }

Sender for a workflow channel provided to the workflow.

A sender is characterized by the type of messages and the codec used to convert messages from / to bytes.

Unlike Receivers, Sender parts of the channel can be cloned. Another difference is ability to control readiness / flushing when sending messages via channel capacity; if the channel reaches its capacity of non-flushed messages, it becomes not ready to accept more messages.

Implementations

impl<T, C: Codec<T>> Sender<T, C>

pub fn closed() -> Self

Returns a sender for a closed channel. This can be provided to spawned workflows to avoid allocating new channels.