Module tardigrade::handle

Handle management for workflows.

Overview

A Handle represents a fundamental building block of a workflow Interface, a receiver or a sender. Depending on the context, receivers and senders may be represented by differing data types (e.g., a channel ID in the host runtime, or a Stream / Sink implementation on the client). Hence, Handle is parameterized by receiver and sender types.

Handles can be collected into a HandleMap, in which they are keyed by a segmented HandlePath. A path is similar to to a path in a filesystem, and just like Path from the standard library, it has an owned counterpart, HandlePathBuf. The segmented nature of paths makes interfaces more composable.

Fallible random access to handles in HandleMaps and other collections is enabled via the HandleMapKey trait. The AccessError type contains detailed information about access errors, such as the path at which the error occurred and the error kind (missing handle, handle kind mismatch, etc.).

HandleMapKey is implemented for a variety of types, including ReceiverAt / SenderAt wrappers that allow specifying the expected handle kind together with the path. The WithIndexing trait allows converting a HandleMap to the form in which it can be indexed by any HandleMapKey impl.

Structs

AccessError

Errors that can occur when accessing an element of a workflow [Interface].

HandleLocation

Location of a Handle in a map.

HandlePath

Path to a Handle in a map.

HandlePathBuf

Owned version of a HandlePath.

IndexingHandleMap

Wrapper for HandleMap allowing accessing the map elements by HandleMapKey. This wrapper is produced by the WithIndexing trait.

ReceiverAt

Newtype for indexing channel receivers, e.g., in an Interface.

SenderAt

Newtype for indexing channel senders, e.g., in an Interface.

Enums

AccessErrorKind

Kind of an AccessError.

Handle

Generic handle to a fundamental building block of a workflow interface (channel sender or receiver).

Traits

HandleMapKey

Key type for accessing handles in a named collections, such as a HandleMap or an Interface.

WithIndexing

Converts HandleMap to the indexing form. This allows accessing contained Handles by any type implementing HandleMapKey, e.g., SenderAt / ReceiverAt.

Type Definitions

HandleMap

Map of handles keyed by owned handle paths.