Struct tardigrade::handle::HandlePath

pub struct HandlePath<'a> { /* private fields */ }

Path to a Handle in a map.

Conceptually, a path is hierarchical like a path in a file system; consists of zero or more string segments. A path can be represented as a string with the segments separated by /. Hierarchical nature of paths is used to compose simplest workflow handles (channel senders and receivers) into composable high-level components.

Examples

const PATH: HandlePath<'_> = HandlePath::new("some/test").join("path");
assert_eq!(PATH.to_string(), "some/test/path");
assert_eq!(PATH.segments().collect::<Vec<_>>(), ["some", "test", "path"]);

let path_buf: HandlePathBuf = PATH.to_owned();
assert_eq!(path_buf.as_ref(), PATH);
let other_path_buf: HandlePathBuf = "other/path".parse()?;

// One of key path properties is ability to use as keys in hash maps:
let mut map = HashMap::<_, _>::from_iter([
    (path_buf, 555),
    (other_path_buf.clone(), 777),
]);
assert_eq!(map[&PATH], 555);
assert_eq!(map[&other_path_buf], 777);

Implementations

impl HandlePath<'static>

pub const EMPTY: HandlePath<'static> = Self{ inner: Inner::Empty,}

Empty path, i.e., a path with zero segments.

impl<'a> HandlePath<'a>

pub const fn new(s: &'a str) -> HandlePath<'a>

Creates a path from the supplied string.

pub const fn join(&'a self, suffix: &'a str) -> HandlePath<'a>

Appends the suffix to this path and returns the resulting path.

pub fn to_owned(self) -> HandlePathBuf

Converts this path to the owned form.

pub fn is_empty(self) -> bool

Checks whether this path is empty (contains no segments).

pub fn segments(self) -> impl Iterator<Item = &'a str>

Iterates over the segments in this path.