Trait tardigrade::handle::HandleMapKey

pub trait HandleMapKey: Into<HandleLocation> + Copy + Debug {
    type Output<Rx, Sx>
   where
        Rx: ?Sized,
        Sx: ?Sized;

    fn get_from<Rx, Sx>(
        self,
        map: &HashMap<HandlePathBuf, Handle<Rx, Sx>, BuildHasherDefault<AHasher>, Global>
    ) -> Result<&Self::Output<Rx, Sx>, AccessError> { ... }
    fn get_mut_from<Rx, Sx>(
        self,
        map: &mut HashMap<HandlePathBuf, Handle<Rx, Sx>, BuildHasherDefault<AHasher>, Global>
    ) -> Result<&mut Self::Output<Rx, Sx>, AccessError> { ... }
    fn remove_from<Rx, Sx>(
        self,
        map: &mut HashMap<HandlePathBuf, Handle<Rx, Sx>, BuildHasherDefault<AHasher>, Global>
    ) -> Result<Self::Output<Rx, Sx>, AccessError> { ... }
}

Key type for accessing handles in a named collections, such as a HandleMap or an Interface.

The methods of this trait rarely need to be used directly; rather, you’d see HandleMapKey as a bound on the access methods of handle collections, e.g. Interface::handle().

Examples

let mut map: HandleMap<u64> = HandleMap::from_iter([
    ("sx".into(), Handle::Sender(42)),
    ("rx".into(), Handle::Receiver(555)),
]);

assert_eq!("sx".get_from(&map)?, &Handle::Sender(42));
*ReceiverAt("rx").get_mut_from(&mut map)? = 777;
assert_eq!(SenderAt("sx").remove_from(&mut map)?, 42);

let err = SenderAt("rx").get_from(&map).unwrap_err();
assert!(matches!(err.kind(), AccessErrorKind::KindMismatch));

Required Associated Types

type Output<Rx, Sx>
where
Rx: ?Sized,
Sx: ?Sized

Output of the access operation. Parameterized by receiver and sender types (e.g., the corresponding specifications for Interface).

Provided Methods

fn get_from<Rx, Sx>(
self,
map: &HashMap<HandlePathBuf, Handle<Rx, Sx>, BuildHasherDefault<AHasher>, Global>
) -> Result<&Self::Output<Rx, Sx>, AccessError>

Returns a shared reference from the provided map using this key.

Errors

Returns an error if a value with the specified key doesn’t exist in the map, or if it has an unexpected type.

fn get_mut_from<Rx, Sx>(
self,
map: &mut HashMap<HandlePathBuf, Handle<Rx, Sx>, BuildHasherDefault<AHasher>, Global>
) -> Result<&mut Self::Output<Rx, Sx>, AccessError>

Returns an exclusive reference from the provided map using this key.

Errors

Returns an error if a value with the specified key doesn’t exist in the map, or if it has an unexpected type.

fn remove_from<Rx, Sx>(
self,
map: &mut HashMap<HandlePathBuf, Handle<Rx, Sx>, BuildHasherDefault<AHasher>, Global>
) -> Result<Self::Output<Rx, Sx>, AccessError>

Removes a value from the provided map using this key.

Errors

Returns an error if a value with the specified key doesn’t exist in the map, or if it has an unexpected type. In the latter case, the value is not removed from the map (i.e., the type check is performed before any modifications).