Struct tardigrade::interface::Interface

pub struct Interface { /* private fields */ }

Specification of a workflow interface. Contains info about channel senders / receivers, arguments etc.

Examples

let interface: Interface = // ...

let spec = interface.handle("commands").unwrap();
println!("{spec:?}");

for (path, spec) in interface.handles() {
    println!("{path}: {spec:?}");
}
// Indexing is also possible using newtype wrappers from the module
let commands = &interface[ReceiverAt("commands")];
println!("{}", commands.description);

Implementations

impl Interface

pub fn try_from_bytes(bytes: &[u8]) -> Result<Interface, Error>

Parses interface definition from bytes.

Currently, this assumes that the definition is JSON-encoded, but this should be considered an implementation detail.

Errors

Returns an error if bytes do not represent a valid interface definition.

pub fn from_bytes(bytes: &[u8]) -> Interface

Version of Self::try_from_bytes() that panics on error.

Panics

Panics if bytes do not represent a valid interface definition.

pub fn to_bytes(&self) -> Vec<u8, Global>

Serializes this interface.

pub fn handle<K>( &self, key: K ) -> Result<&<K as HandleMapKey>::Output<ReceiverSpec, SenderSpec>, AccessError>where K: HandleMapKey,

Returns spec for a handle defined in this interface.

Errors

Returns an error if a spec with the specified key doesn’t exist in the interfaces, or if it has an unexpected type.

pub fn handles(&self) -> impl ExactSizeIterator

Lists all handle specifications in this interface.

pub fn args(&self) -> &ArgsSpec

Returns spec for the arguments.

pub fn check_compatibility( &self, provided: &Interface ) -> Result<(), AccessError>

Checks the compatibility of this expected interface against the provided interface. The provided interface may contain more channels than is described by the expected interface, but not vice versa.

Errors

Returns an error if the provided interface does not match expectations.

pub fn check_shape<Rx, Sx>( &self, handles: &HashMap<HandlePathBuf, Handle<Rx, Sx>, BuildHasherDefault<AHasher>, Global>, exact: bool ) -> Result<(), AccessError>

Checks that the shape of the provided handles corresponds to this interface. If the exact flag is set, the shape match must be exact; otherwise, there might be extra handles not specified by this interface.

Errors

Returns an error on shape mismatch.