Type Alias tardigrade::interface::HandleSpec

pub type HandleSpec = Handle<ReceiverSpec, SenderSpec>;

Specification for a handle in a workflow Interface.

Aliased Type

enum HandleSpec {
    Receiver(ReceiverSpec),
    Sender(SenderSpec),
}

Variants

Receiver(ReceiverSpec)

Receiver handle.

Sender(SenderSpec)

Sender handle.

Implementations

impl<Rx, Sx> Handle<Rx, Sx>

pub fn as_ref(&self) -> Handle<&Rx, &Sx>

Converts a shared reference to this handle to a Handle containing the shared reference.

pub fn as_mut(&mut self) -> Handle<&mut Rx, &mut Sx>

Converts an exclusive reference to this handle to a Handle containing the exclusive reference.

pub fn map_receiver<U>(self, mapping: impl FnOnce(Rx) -> U) -> Handle<U, Sx>

Maps the receiver type in this handle.

pub fn map_sender<U>(self, mapping: impl FnOnce(Sx) -> U) -> Handle<Rx, U>

Maps the sender type of this handle.

impl<T> Handle<T, T>

pub fn map<U>(self, mapping: impl FnOnce(T) -> U) -> Handle<U, U>

Maps both types of this handle provided that they coincide.

pub fn factor(self) -> T

Factors the underlying type from this handle.