Module tardigrade::spawn

Spawning and managing child workflows.

Examples

// Assume we want to spawn a child workflow defined as follows:
#[derive(GetInterface, WithHandle)]
#[tardigrade(derive(Debug), auto_interface)]
pub struct ChildWorkflow<Env: HandleFormat = Wasm> {
    pub commands: InEnv<Receiver<String, Json>, Env>,
    pub events: InEnv<Sender<String, Json>, Env>,
}

impl WorkflowFn for ChildWorkflow {
    type Args = ();
    type Codec = Json;
}

// To spawn a workflow, we should use the following code
// in the parent workflow:
use tardigrade::spawn::{CreateWorkflow, Workflows};

let builder = Workflows.new_workflow::<ChildWorkflow>("child")?;
let (child_handles, mut self_handles) = builder
    .handles(|config| {
        // It is possible to customize child workflow initialization via
        // `builder.handle()`, but zero config is fine as well.
    })
    .await;
let child = builder.build((), child_handles).await?;
// `child` contains handles to the created channels.
self_handles.commands.send("ping".to_owned()).await?;
let event: String = self_handles.events.next().await.unwrap();
child.await?;

Structs

HostError

Errors generated on the host side and sent to a workflow.

ReceiverConfig

Configurator of a workflow channel Receiver.

RemoteWorkflow

Handle to a remote workflow (usually, a child workflow previously spawned using Workflows).

SenderConfig

Configurator of a workflow channel Sender.

Spawner

Spawn environment that can be used to configure channels before spawning a workflow.

WorkflowBuilder

Builder of child workflows.

Workflows

Client-side connection to a [workflow manager][ManageWorkflows].

Traits

CreateChannel

Manager of workflow channels.

CreateWorkflow

Manager of workflows that allows spawning them.

ManageInterfaces

Manager of Interfaces that allows obtaining an interface by a string identifier.

Type Definitions

HandlesPair

Pair of handles with a certain interface in a certain format and the inverse format.