Trait tardigrade::spawn::CreateWorkflow

pub trait CreateWorkflow: ManageInterfaces + CreateChannel {
    type Spawned<W: WorkflowFn + WithHandle>;
    type Error: 'static + Send + Sync;

    fn new_workflow<'a, W: WorkflowFn + GetInterface>(
        &'a self,
        definition_id: &'a str
    ) -> Result<WorkflowBuilder<'a, Self, W>, AccessError>
   where
        Self: Sized,
    { ... }
}

Manager of workflows that allows spawning them.

Depending on the context, the spawned workflow may be a child workflow (if executed in the context of a workflow, via Workflows), or the top-level workflow (if executed from the host).

Required Associated Types

type Spawned<W: WorkflowFn + WithHandle>

Handle to an instantiated workflow.

type Error: 'static + Send + Sync

Error spawning a workflow.

Provided Methods

fn new_workflow<'a, W: WorkflowFn + GetInterface>(
&'a self,
definition_id: &'a str
) -> Result<WorkflowBuilder<'a, Self, W>, AccessError>where
Self: Sized,

Initiates creating a new workflow and returns the corresponding builder.

Errors

Returns an error on interface mismatch between W and the workflow definition contained in this manager under definition_id.