Trait tardigrade::spawn::ManageInterfaces

pub trait ManageInterfaces {
    fn interface(&self, definition_id: &str) -> Option<Cow<'_, Interface>>;
}

Manager of Interfaces that allows obtaining an interface by a string identifier.

Required Methods

fn interface(&self, definition_id: &str) -> Option<Cow<'_, Interface>>

Returns the interface spec of a workflow with the specified ID.