Module tardigrade::task

Task management in workflows.

Similarly to other async runtimes, a task is a self-contained unit of work with the 'static lifetime. Tasks are spawned from the workflow code using spawn() or try_spawn() functions, which return the JoinHandle for the created task. This handle can be polled to completion, or be used to abort the task.

Error handling

Errors during task execution are encapsulated in TaskError, which provides a way to track source code location and the cause of errors. Unlike with threads or tasks in most other async runtimes, a panic in any task is unrecoverable. Depending on the host settings, it may lead to the erroneous workflow termination, or to its rollback.

Structs

ErrorContext

Additional context for a TaskError.

ErrorLocation

Location of a panic or error in the workflow code.

JoinHandle

Handle to a spawned task.

TaskError

Business error raised by a workflow task.

Enums

JoinError

Errors that can occur when joining a task or a previously spawned workflow (i.e., waiting for its completion).

Traits

ErrorContextExt

Extension trait to convert Results and Options to TaskResults.

Functions

spawn

Spawns a new infallible task and returns a handle that can be used to wait for its completion or abort the task.

try_spawn

Spawns a new fallible task and returns a handle that can be used to wait for its completion or abort the task.

yield_now

Yields execution of the current task, allowing to switch to other tasks.

Type Definitions

TaskResult

Result of executing a workflow task.