Module tardigrade::test

Utils for workflow testing.

This emulates imports / exports provided to WASM workflow modules by the Tardigrade runtime, allowing to test workflow logic without launching a runtime (or even including the corresponding crate as a dev dependency).

Examples

// Assume we want to test this workflow.
#[derive(WithHandle, GetInterface)]
#[tardigrade(derive(Debug), auto_interface)]
pub struct MyWorkflow<Fmt: HandleFormat = Wasm> {
    pub events: InEnv<Sender<Event, Json>, Fmt>,
}

/// Arguments provided to the workflow on creation.
#[derive(Debug, Serialize, Deserialize)]
pub struct Args {
    pub counter: u32,
}

/// Events emitted by the workflow.
#[derive(Debug, Serialize, Deserialize)]
pub enum Event {
    Count(u32),
}

impl WorkflowFn for MyWorkflow {
    type Args = Args;
    type Codec = Json;
}

// Workflow logic
#[async_trait(?Send)]
impl SpawnWorkflow for MyWorkflow {
    async fn spawn(args: Args, mut handle: Self) -> TaskResult {
        for i in 0..args.counter {
            tardigrade::sleep(Duration::from_millis(100)).await;
            handle.events.send(Event::Count(i)).await.ok();
        }
        Ok(())
    }
}

// We can test the workflow as follows:
use tardigrade::{spawn::RemoteWorkflow, test::{TestInstance, Timers}};

async fn test_workflow(handle: MyWorkflow<Inverse<Wasm>>) {
    // The workflow should be waiting for a timer to emit an event.
    let mut events = handle.events;
    assert!(events.next().now_or_never().is_none());

    let now = Timers::now();
    let new_now = Timers::next_timer_expiration().unwrap();
    assert_eq!((new_now - now).num_milliseconds(), 100);
    Timers::set_now(new_now);
    let event = events.next().await.unwrap();
    assert_matches!(event, Event::Count(0));
}
TestInstance::<MyWorkflow>::new(Args { counter: 1 }).run(test_workflow);

Structs

MockScheduler

Mock workflow scheduler.

Runtime

Mock runtime allowing to execute workflows.

TestInstance

Instance of a workflow combined with a Runtime.

Timers

Handle allowing to manipulate time in the test environment.

WorkflowRegistry

Mock registry for workflows.