Struct tardigrade::test::Runtime

pub struct Runtime { /* private fields */ }

Mock runtime allowing to execute workflows.

Implementations

impl Runtime

pub fn workflow_registry_mut(&mut self) -> &mut WorkflowRegistry

Returns a mutable reference to the workflow registry.

pub fn run<Fut: Future>(self, test_future: Fut) -> Fut::Output

Executes the provided future in the context of this runtime. Thus, the future will have access to globals provided by a runtime such as Workflows, or test-specific globals such as Timers.

This method can be used to handle more complex tests than Self::test(), e.g., ones that do not want to spin a specific workflow.

pub fn test<W>(self, args: W::Args) -> TestInstance<W>where
W: SpawnWorkflow,

Tests the specified workflow definition in the context of this runtime.