Trait tardigrade::Codec

pub trait Codec<T>: 'static + Send + Sync {
    type Error: Error + Send + Sync + 'static;

    fn encode_value(value: T) -> Vec<u8>;
    fn try_decode_bytes(bytes: Vec<u8>) -> Result<T, Self::Error>;

    fn decode_bytes(bytes: Vec<u8>) -> T { ... }
}

Codec for a particular data type.

When used for decoding in workflows (e.g., in Receiver), Self::decode_bytes() method is used. That is, decoding operations panic if an error occurs during decoding. This is usually the intended behavior, since the primary reason of such a panic would be an incorrect input supplied via arguments or a channel receiver. The workflow runtime is expected to roll back the workflow state in this case.

Required Associated Types

type Error: Error + Send + Sync + 'static

Decoding error.

Required Methods

fn encode_value(value: T) -> Vec<u8>

Encodes value.

fn try_decode_bytes(bytes: Vec<u8>) -> Result<T, Self::Error>

Tries to decode bytes.

Errors

Returns an error if bytes do not represent valid data.

Provided Methods

fn decode_bytes(bytes: Vec<u8>) -> T

Decodes bytes. This is a convenience method that unwrap()s the result of Self::try_decode_bytes().

Panics

Panics if bytes do not represent valid data.