Module tardigrade::workflow

Workflow-related types.

See the crate docs for an intro on workflows.

Examples

Simple workflow definition:

use tardigrade::{
    channel::{Sender, Receiver}, task::TaskResult, workflow::*, Json,
};

/// Handle for the workflow. Fields are public for integration testing.
#[derive(WithHandle, GetInterface, WorkflowEntry)]
#[tardigrade(derive(Debug), auto_interface)]
pub struct MyWorkflow<Fmt: HandleFormat = Wasm> {
    /// Receiver for commands.
    pub commands: InEnv<Receiver<Command, Json>, Fmt>,
    /// Sender for events.
    pub events: InEnv<Sender<Event, Json>, Fmt>,
}

/// Args provided to the workflow on creation.
#[derive(Debug, Serialize, Deserialize)]
pub struct Args {
    pub start_counter: u32,
}

/// Commands received via `commands` channel.
#[derive(Debug, Serialize, Deserialize)]
pub enum Command {
    Ping(String),
    // other variants...
}

/// Events emitted via `events` channel.
#[derive(Debug, Serialize, Deserialize)]
pub enum Event {
    Pong(String),
    // other variants...
}

impl MyWorkflow {
    async fn process_command(
        &mut self,
        command: &Command,
        counter: &mut u32,
    ) {
        match command {
            Command::Ping(ping) => {
                let pong = format!("{ping}, counter={}", *counter);
                *counter += 1;
                self.events.send(Event::Pong(pong)).await.ok();
            }
            // other commands...
        }
    }
}

// Workflow interface declaration.
impl WorkflowFn for MyWorkflow {
    type Args = Args;
    type Codec = Json;
}

// Actual workflow logic.
#[async_trait(?Send)]
impl SpawnWorkflow for MyWorkflow {
    async fn spawn(args: Args, mut handle: Self) -> TaskResult {
        let mut counter = args.start_counter;
        while let Some(command) = handle.commands.next().await {
            handle
                .process_command(&command, &mut counter)
                .await;
        }
        Ok(())
    }
}

Structs

Inverse

Wrapper for a HandleFormat that swaps handles for senders and receivers.

Wasm

WASM environment.

Traits

DelegateHandle

Implementation of WithHandle that delegates to the provided type. This is useful for sharing handles among multiple workflows.

GetInterface

Allows obtaining an Interface for a workflow.

HandleFormat

Format of Handles (senders and receivers).

InsertHandles

Accumulator of handles in a certain format used in WithHandle::insert_into_untyped().

IntoRaw

Conversion into a raw handle.

SpawnWorkflow

Workflow that can be spawned.

TakeHandles

Collection of handles in a certain format that can be taken from. Used in WithHandle::take_from_untyped().

TryFromRaw

Conversion from a raw handle.

WithHandle

Type with a polymorphic handle depending on a HandleFormat.

WorkflowEntry

Workflow that is accessible by its name from a module.

WorkflowFn

Functional interface of a workflow.

Type Definitions

InEnv

Handle in a particular format.

UntypedHandles

Dynamically-typed handle to a workflow containing handles to its channels.

Derive Macros

GetInterface

Derives the GetInterface trait for a workflow type.

WithHandle

Derives the WithHandle trait for a workflow type.

WorkflowEntry

Derives the WorkflowEntry trait for a workflow type.