Trait tardigrade::workflow::GetInterface

pub trait GetInterface: WithHandle + Sized + 'static {
    fn interface() -> Cow<'static, Interface> { ... }
}

Allows obtaining an Interface for a workflow.

This trait should be derived for workflow types using the corresponding macro.

Provided Methods

fn interface() -> Cow<'static, Interface>

Obtains the workflow interface.

The default implementation uses the WithHandle implementation to create an owned interface. The GetInterface derive macro provides a more efficient cached implementation.