Trait tardigrade::workflow::HandleFormat

pub trait HandleFormat {
    type RawReceiver;
    type Receiver<T, C: Codec<T>>: IntoRaw<Self::RawReceiver> + TryFromRaw<Self::RawReceiver>;
    type RawSender;
    type Sender<T, C: Codec<T>>: IntoRaw<Self::RawSender> + TryFromRaw<Self::RawSender>;
}

Format of Handles (senders and receivers).

The format specifies how receivers and senders are represented in terms of data types, and how they are converted to / from their raw representations.

Required Associated Types

type RawReceiver

Raw receiver handle.

type Receiver<T, C: Codec<T>>: IntoRaw<Self::RawReceiver> + TryFromRaw<Self::RawReceiver>

Receiver handle in this format. Type parameters in the handle is the received value and its codec, respectively.

type RawSender

Raw sender handle.

type Sender<T, C: Codec<T>>: IntoRaw<Self::RawSender> + TryFromRaw<Self::RawSender>

Sender handle in this format. Type parameters in the handle is the received value and its codec, respectively.