Trait tardigrade::workflow::TakeHandles

pub trait TakeHandles<Fmt: HandleFormat> {
    fn take_receiver(
        &mut self,
        path: HandlePath<'_>
    ) -> Result<Fmt::RawReceiver, AccessError>;
    fn take_sender(
        &mut self,
        path: HandlePath<'_>
    ) -> Result<Fmt::RawSender, AccessError>;
    fn drain(&mut self) -> UntypedHandles<Fmt>;
}

Collection of handles in a certain format that can be taken from. Used in WithHandle::take_from_untyped().

Required Methods

fn take_receiver(
&mut self,
path: HandlePath<'_>
) -> Result<Fmt::RawReceiver, AccessError>

Takes a raw receiver handle at the specified path from this collection.

Errors

Returns an error if there is no receiver with the specified path.

fn take_sender(
&mut self,
path: HandlePath<'_>
) -> Result<Fmt::RawSender, AccessError>

Takes a raw sender handle at the specified path from this collection.

Errors

Returns an error if there is no sender with the specified path.

fn drain(&mut self) -> UntypedHandles<Fmt>

Drains this collection into untyped handles.