Crate tardigrade

Client bindings for Tardigrade workflows. This crate should be used by WASM modules defining a workflow.

What’s a workflow, anyway?

A workflow is a essentially a Future with () output that interacts with the external world via well-defined interfaces:

Arguments provided to the workflow on creation

Message channels similar to ones from the future crate

Timers

Tasks

Timers can be created dynamically during workflow operation; likewise, tasks can be spawn()ed to achieve concurrency (but not parallelism!). In contrast, the set of channel halves and their direction (receiver or sender) are static / predefined in the workflow Interface.

A workflow is sandboxed by the virtue of being implemented as a WASM module. This brings multiple benefits:

A workflow can be suspended, its state persisted and then restored completely transparently for workflow logic. Further, this approach does not restrict workflow functionality; it can do essentially anything Futures can do: fork / joins, fork / selects, retries etc.

The runtime can (theoretically) customize aspects of workflow execution, e.g., priorities of different tasks. It can create and escalate incidents on task failure (perhaps, continuing executing other tasks if possible), control rollback of the workflow state etc.

A workflow could (theoretically) implemented in any WASM-targeting programming language and can run / be debugged in any WASM-supporting env (e.g., in a browser).

Sandboxing means that interfaces for interacting with the external world are well-defined (e.g., in terms of security) and can be emulated (e.g., for testing) with reasonable effort.

Crate features

tracing

(On by default)

Enables tracing for the library glue code and the workflows that are defined using the library.

Modules

channel

Message channels for workflows.

handle

Handle management for workflows.

interface

Types related to workflow interface definition.

spawn

Spawning and managing child workflows.

task

Task management in workflows.

test

Utils for workflow testing.

workflow

Workflow-related types.

Structs

Json

JSON codec.

Raw

Raw / identity codec that passes through byte Vecs without changes.

Timer

A future emitting time events.

TimerDefinition

Definition of a timer used by a workflow.

Traits

Codec

Codec for a particular data type.

Functions

now

Returns the current timestamp as per the wall clock associated with the workflow.

sleep

Sleeps for the specified duration.

Type Aliases

ChannelId

ID of a channel.

TaskId

ID of a workflow task.

TimerId

ID of a workflow timer.

WakerId

ID of a Waker defined by a workflow.

WorkflowId

ID of a workflow.