Trait tardigrade::workflow::WithHandle

pub trait WithHandle {
    type Handle<Fmt: HandleFormat>;

    // Required methods
    fn take_from_untyped<Fmt: HandleFormat>(
        untyped: &mut impl TakeHandles<Fmt>,
        path: HandlePath<'_>
    ) -> Result<Self::Handle<Fmt>, AccessError>;
    fn insert_into_untyped<Fmt: HandleFormat>(
        handle: Self::Handle<Fmt>,
        untyped: &mut impl InsertHandles<Fmt>,
        path: HandlePath<'_>
    );

    // Provided methods
    fn try_from_untyped<Fmt: HandleFormat>(
        untyped: UntypedHandles<Fmt>
    ) -> Result<Self::Handle<Fmt>, AccessError> { ... }
    fn into_untyped<Fmt: HandleFormat>(
        handle: Self::Handle<Fmt>
    ) -> UntypedHandles<Fmt> { ... }
}

Type with a polymorphic handle depending on a HandleFormat.

Conceptually, a handle collects simplest Handles (that for receivers and senders) into more complex structures. A handle can be taken from an untyped collection of handles (e.g., UntypedHandles) or, conversely, inserted into an untyped collection.

A handle needs to be defined for a workflow in order to specify which handles it receives on initialization. The intended way to do this is to define a dedicated handle type with the help of the WithHandle derive macro. If necessary, its implementation can be delegated from the (usually empty) workflow type. See the workflow module docs for an example.

Required Associated Types

type Handle<Fmt: HandleFormat>

Type of the handle in a particular format.

Required Methods

fn take_from_untyped<Fmt: HandleFormat>( untyped: &mut impl TakeHandles<Fmt>, path: HandlePath<'_> ) -> Result<Self::Handle<Fmt>, AccessError>

Produces the handle from an untyped collection of handles.

Errors

Returns an error if the collection does not have the necessary shape.

fn insert_into_untyped<Fmt: HandleFormat>( handle: Self::Handle<Fmt>, untyped: &mut impl InsertHandles<Fmt>, path: HandlePath<'_> )

Inserts the handle into an untyped accumulator.

Provided Methods

fn try_from_untyped<Fmt: HandleFormat>( untyped: UntypedHandles<Fmt> ) -> Result<Self::Handle<Fmt>, AccessError>

Produces the handle from an untyped collection of handles. This is a higher-level alternative to Self::take_from_untyped().

Errors

Returns an error in the same situations as Self::take_from_untyped().

fn into_untyped<Fmt: HandleFormat>( handle: Self::Handle<Fmt> ) -> UntypedHandles<Fmt>

Converts this handle into UntypedHandles. This is a higher-level alternative to Self::insert_into_untyped().